break up main applet code into multiple files for readability
