+++ software/obi/applet/open_beam_interface/__init__.py
@@ -12,14 +12,14 @@ from amaranth.lib.wiring import In, Out, flipped
 from glasgow.applet import GlasgowApplet
 from glasgow.support.logging import dump_hex
 from glasgow.support.endpoint import ServerEndpoint
-from glasgow.device import GlasgowDeviceError
 
-from usb1 import USBError
-
-from obi.commands.structs import CmdType, BeamType, OutputMode, Transforms
-from obi.commands.low_level_commands import Command, ExternalCtrlCommand
-
-import logging
+from obi.applet.open_beam_interface.modules.structs import Transforms
+from obi.commands import (CmdType, BeamType, OutputMode, 
+                        SynchronizeCommand, FlushCommand, VectorPixelMinDwellCommand)
+from obi.applet.open_beam_interface.modules import (
+    Transforms, BlankRequest,BusSignature, DwellTime, DACStream, SuperDACStream, 
+    PipelinedLoopbackAdapter, BusController, FastBusController, 
+    Supersampler, RasterScanner, CommandParser)
 
 # Overview of (linear) processing pipeline:
 # 1. PC software (in: user input, out: bytes)
@@ -34,659 +34,6 @@ import logging
 # 10. Frame serializer (in: frames, out: bytes)
 # 11. Glasgow software/framework (in: bytes, out: same bytes; vendor-provided)
 # 12. PC software (in: bytes, out: displayed image)
-
-
-def StreamSignature(data_layout):
-    return wiring.Signature({
-        "payload":  Out(data_layout),
-        "valid": Out(1),
-        "ready": In(1)
-    })
-
-class BlankRequest(data.Struct):
-    enable: 1
-    request: 1
-
-
-#=========================================================================
-
-class SkidBuffer(wiring.Component):
-    def __init__(self, data_layout, *, depth):
-        self.width = Shape.cast(data_layout).width
-        self.depth = depth
-        super().__init__({
-            "i": In(StreamSignature(data_layout)),
-            "o": Out(StreamSignature(data_layout)),
-        })
-
-    def elaborate(self, platform):
-        m = Module()
-
-        m.submodules.fifo = fifo = \
-            SyncFIFOBuffered(depth=self.depth, width=self.width)
-        m.d.comb += [
-            fifo.w_data.eq(self.i.payload),
-            fifo.w_en.eq(self.i.valid),
-            self.i.ready.eq(fifo.level <= 1),
-            self.o.payload.eq(fifo.r_data),
-            self.o.valid.eq(fifo.r_rdy),
-            fifo.r_en.eq(self.o.ready),
-        ]
-
-        return m
-
-#=========================================================================
-BusSignature = wiring.Signature({
-    "adc_clk":  Out(1),
-    "adc_le_clk":   Out(1),
-    "adc_oe":   Out(1),
-
-    "dac_clk":  Out(1),
-    "dac_x_le_clk": Out(1),
-    "dac_y_le_clk": Out(1),
-
-    "data_i":   In(15),
-    "data_o":   Out(15),
-    "data_oe":  Out(1),
-})
-
-DwellTime = unsigned(16)
-
-class PipelinedLoopbackAdapter(wiring.Component):
-    loopback_stream: In(unsigned(14))
-    bus: Out(BusSignature)
-
-    def __init__(self, adc_latency: int):
-        self.adc_latency = adc_latency
-        super().__init__()
-
-    def elaborate(self, platform):
-        m = Module()
-
-        prev_bus_adc_oe = Signal()
-        adc_oe_falling = Signal()
-        m.d.sync += prev_bus_adc_oe.eq(self.bus.adc_oe)
-        m.d.comb += adc_oe_falling.eq(prev_bus_adc_oe & ~self.bus.adc_oe)
-
-        shift_register = Signal(14*self.adc_latency)
-
-        with m.If(adc_oe_falling):
-            m.d.sync += shift_register.eq((shift_register << 14) | self.loopback_stream)
-
-        m.d.comb += self.bus.data_i.eq(shift_register.word_select(self.adc_latency-1, 14))
-
-        return m
-
-
-class DACStream(data.Struct):
-    dac_x_code: 14
-    padding_x: 2
-    dac_y_code: 14
-    padding_x: 2
-    dwell_time: 16
-    blank: BlankRequest
-    delay: 3
-
-
-class SuperDACStream(data.Struct):
-    dac_x_code: 14
-    padding_x: 2
-    dac_y_code: 14
-    padding_y: 2
-    blank: BlankRequest
-    last:       1
-    delay: 3
-
-
-
-class BusController(wiring.Component):
-    # FPGA-side interface
-    dac_stream: In(StreamSignature(SuperDACStream))
-
-    adc_stream: Out(StreamSignature(data.StructLayout({
-        "adc_code": 14,
-        "adc_ovf":  1,
-        "last":     1,
-    })))
-
-    # IO-side interface
-    bus: Out(BusSignature)
-    inline_blank: Out(BlankRequest)
-
-    def __init__(self, *, adc_half_period: int, adc_latency: int, transforms: Transforms = Transforms(False,False,False)):
-        assert (adc_half_period * 2) >= 6, "ADC period must be large enough for FSM latency"
-        self.adc_half_period = adc_half_period
-        self.adc_latency     = adc_latency
-        self.transforms = transforms
-
-        super().__init__()
-
-        self.dac_x_code_transformed = Signal.like(self.dac_stream.payload.dac_x_code)
-        self.dac_y_code_transformed = Signal.like(self.dac_stream.payload.dac_y_code)
-
-    def elaborate(self, platform):
-        m = Module()
-
-        adc_cycles = Signal(range(self.adc_half_period))
-        with m.If(adc_cycles == self.adc_half_period - 1):
-            m.d.sync += adc_cycles.eq(0)
-            m.d.sync += self.bus.adc_clk.eq(~self.bus.adc_clk)
-        with m.Else():
-            m.d.sync += adc_cycles.eq(adc_cycles + 1)
-        # ADC and DAC share the bus and have to work in tandem. The ADC conversion starts simultaneously
-        # with the DAC update, so the entire ADC period is available for DAC-scope-ADC propagation.
-        m.d.comb += self.bus.dac_clk.eq(self.bus.adc_clk)
-
-
-        # Queue; MSB = most recent sample, LSB = least recent sample
-        accept_sample = Signal(self.adc_latency)
-        # Queue; as above
-        last_sample = Signal(self.adc_latency)
-
-        m.submodules.skid_buffer = skid_buffer = \
-            SkidBuffer(self.adc_stream.payload.shape(), depth=self.adc_latency)
-        wiring.connect(m, flipped(self.adc_stream), skid_buffer.o)
-
-        adc_stream_data = Signal.like(self.adc_stream.payload) # FIXME: will not be needed after FIFOs have shapes
-        m.d.comb += [
-            # Cat(adc_stream_data.adc_code,
-            #     adc_stream_data.adc_ovf).eq(self.bus.i),
-            adc_stream_data.last.eq(last_sample[self.adc_latency-1]),
-            skid_buffer.i.payload.eq(adc_stream_data),
-        ]
-
-        dac_stream_data = Signal.like(self.dac_stream.payload)
-        
-        x = Signal.like(self.dac_x_code_transformed)
-        y = Signal.like(self.dac_y_code_transformed)
-
-        m.d.comb += adc_stream_data.adc_code.eq(self.bus.data_i)
-
-        stalled = Signal()
-
-        with m.FSM():
-            with m.State("ADC_Wait"):
-                with m.If(self.bus.adc_clk & (adc_cycles == 0)):
-                    m.d.comb += self.bus.adc_le_clk.eq(1)
-                    m.d.comb += self.bus.adc_oe.eq(1) #give bus time to stabilize before sampling
-                    m.next = "ADC_Read"
-
-            with m.State("ADC_Read"):
-                #m.d.comb += self.bus.adc_le_clk.eq(1)
-                m.d.comb += self.bus.adc_oe.eq(1)
-                # buffers up to self.adc_latency samples if skid_buffer.i.ready
-                m.d.comb += skid_buffer.i.valid.eq(accept_sample[self.adc_latency-1])
-                with m.If(self.dac_stream.valid & skid_buffer.i.ready):
-                    # Latch DAC codes from input stream.
-                    m.d.comb += self.dac_stream.ready.eq(1)
-                    m.d.sync += dac_stream_data.eq(self.dac_stream.payload)
-                    # Transforms
-                    # Rotate first so that x is x and y is y, then flip x and y as needed
-                    if self.transforms.rotate90:
-                        m.d.comb += x.eq(self.dac_stream.payload.dac_y_code)
-                        m.d.comb += y.eq(self.dac_stream.payload.dac_x_code)
-                    else:
-                        m.d.comb += x.eq(self.dac_stream.payload.dac_x_code)
-                        m.d.comb += y.eq(self.dac_stream.payload.dac_y_code)
-                    
-                    if self.transforms.xflip:
-                        m.d.sync += self.dac_x_code_transformed.eq(16383-x)
-                    else:
-                        m.d.sync += self.dac_x_code_transformed.eq(x)
-
-                    if self.transforms.yflip:
-                        m.d.sync += self.dac_y_code_transformed.eq(16383-y)
-                    else:
-                        m.d.sync += self.dac_y_code_transformed.eq(y)
-
-                    # Transmit blanking state from input stream
-                    m.d.comb += self.inline_blank.eq(self.dac_stream.payload.blank)
-                    # Schedule ADC sample for these DAC codes to be output.
-                    m.d.sync += accept_sample.eq(Cat(1, accept_sample))
-                    # Carry over the flag for last sample [of averaging window] to the output.
-                    m.d.sync += last_sample.eq(Cat(self.dac_stream.payload.last, last_sample))
-                with m.Else():
-                    # Leave DAC codes as they are.
-                    # Schedule ADC sample for these DAC codes to be discarded.
-                    m.d.sync += accept_sample.eq(Cat(0, accept_sample))
-                    # The value of this flag is discarded, so it doesn't matter what it is.
-                    m.d.sync += last_sample.eq(Cat(0, last_sample))
-                m.next = "X_DAC_Write"
-
-            with m.State("X_DAC_Write"):
-                m.d.comb += [
-                    self.bus.data_o.eq(self.dac_x_code_transformed),
-                    self.bus.data_oe.eq(1),
-                ]
-                m.next = "X_DAC_Write_2"
-
-            with m.State("X_DAC_Write_2"):
-                m.d.comb += [
-                    self.bus.data_o.eq(self.dac_x_code_transformed),
-                    self.bus.data_oe.eq(1),
-                    self.bus.dac_x_le_clk.eq(1),
-                ]
-                m.next = "Y_DAC_Write"
-
-            with m.State("Y_DAC_Write"):
-                m.d.comb += [
-                    self.bus.data_o.eq(self.dac_y_code_transformed),
-                    self.bus.data_oe.eq(1),
-                ]
-                m.next = "Y_DAC_Write_2"
-
-            with m.State("Y_DAC_Write_2"):
-                m.d.comb += [
-                    self.bus.data_o.eq(self.dac_y_code_transformed),
-                    self.bus.data_oe.eq(1),
-                    self.bus.dac_y_le_clk.eq(1),
-                ]
-                m.next = "ADC_Wait"
-
-        return m
-
-class FastBusController(wiring.Component):
-    # FPGA-side interface
-    dac_stream: In(StreamSignature(SuperDACStream))
-
-
-    # IO-side interface
-    bus: Out(BusSignature)
-    inline_blank: Out(BlankRequest)
-
-    # Ignored
-    adc_stream: Out(StreamSignature(data.StructLayout({
-        "adc_code": 14,
-        "adc_ovf":  1,
-        "last":     1,
-    })))
-
-    def __init__(self):
-        self.delay = 3
-        super().__init__()
-
-    def elaborate(self, platform):
-        m = Module()
-
-        delay_cycles = Signal(3)
-
-        dac_stream_data = Signal.like(self.dac_stream.payload)
-        m.d.comb += self.inline_blank.eq(dac_stream_data.blank)
-
-        with m.FSM():
-            with m.State("X_DAC_Write"):
-                m.d.comb += [
-                    self.bus.data_o.eq(dac_stream_data.dac_x_code),
-                    self.bus.data_oe.eq(1),
-                ]
-                m.d.sync += self.bus.dac_clk.eq(0)
-                m.next = "X_DAC_Write_2"
-
-            with m.State("X_DAC_Write_2"):
-                m.d.comb += [
-                    self.bus.data_o.eq(dac_stream_data.dac_x_code),
-                    self.bus.data_oe.eq(1),
-                    self.bus.dac_x_le_clk.eq(1),
-                ]
-                m.next = "Y_DAC_Write"
-
-            with m.State("Y_DAC_Write"):
-                m.d.comb += [
-                    self.bus.data_o.eq(dac_stream_data.dac_y_code),
-                    self.bus.data_oe.eq(1),
-                ]
-                m.next = "Y_DAC_Write_2"
-
-            with m.State("Y_DAC_Write_2"):
-                m.d.comb += [
-                    self.bus.data_o.eq(dac_stream_data.dac_y_code),
-                    self.bus.data_oe.eq(1),
-                    self.bus.dac_y_le_clk.eq(1),
-                ]  
-                m.next = "Latch_Delay"
-            
-            with m.State("Latch_Delay"):
-                with m.If(delay_cycles > 0):
-                    m.d.sync += delay_cycles.eq(delay_cycles - 1) 
-                with m.Else():
-                    m.d.sync += self.bus.dac_clk.eq(1)
-                    with m.If(self.dac_stream.valid):
-                        # Latch DAC codes from input stream.
-                        m.d.comb += self.dac_stream.ready.eq(1)
-                        m.d.sync += dac_stream_data.eq(self.dac_stream.payload)
-                        with m.If(dac_stream_data.last): #latch delay from the previous stream
-                            m.d.sync +=  delay_cycles.eq(dac_stream_data.delay)
-                    m.next = "X_DAC_Write"  
-                
-                    
-
-        return m
-
-#=========================================================================
-
-class PowerOfTwoDetector(Elaboratable):
-    """Priority encode requests to binary.
-
-    If any bit in ``i`` is asserted, ``n`` is low and ``o`` indicates the least significant
-    asserted bit.
-    Otherwise, ``n`` is high and ``o`` is ``0``.
-
-    Parameters
-    ----------
-    width : int
-        Bit width of the input.
-
-    Attributes
-    ----------
-    i : Signal(width), in
-        Input requests.
-    o : Signal(range(width)), out
-        Encoded natural binary.
-    n : Signal, out
-        Invalid: no input bits are asserted.
-    """
-    def __init__(self, width):
-        self.width = width
-
-        self.i = Signal(width)
-        self.o = Signal(range(width))
-        self.n = Signal()
-        self.p = Signal()
-
-    def elaborate(self, platform):
-        m = Module()
-        p = Signal()
-        for power in range(self.width):
-            with m.If(self.i[power]):
-                m.d.comb += self.o.eq(power)
-            with m.If(self.i == 1 << power):
-                m.d.comb += p.eq(1)
-        m.d.comb += self.n.eq(self.i == 0)
-        m.d.comb += self.p.eq(p & ~self.n)
-        return m
-
-
-class Supersampler(wiring.Component):
-    """
-    In:
-        dac_stream: X and Y DAC codes and dwell time
-        super_adc_stream: ADC sample value and `last` signal
-    Out:
-        super_dac_stream: X and Y DAC codes and `last` signal
-        adc_stream: Averaged ADC sample value
-    """
-    dac_stream: In(StreamSignature(DACStream))
-
-    adc_stream: Out(StreamSignature(data.StructLayout({
-        "adc_code":   14,
-    })))
-
-    super_dac_stream: Out(StreamSignature(SuperDACStream))
-
-    super_adc_stream: In(StreamSignature(data.StructLayout({
-        "adc_code":   14,
-        "adc_ovf":    1,  # ignored
-        "last":       1,
-    })))
-
-    ## debug info
-    stall_cycles: Out(16)
-    stall_count_reset: In(1)
-
-    def __init__(self):
-        super().__init__()
-
-        self.dac_stream_data = Signal.like(self.dac_stream.payload)
-        self.encoder = PowerOfTwoDetector(16)
-
-    def elaborate(self, platform):
-        m = Module()
-        m.submodules["encoder"] = self.encoder
-
-        dwell_counter = Signal.like(self.dac_stream_data.dwell_time)
-        sample_counter = Signal.like(self.dac_stream_data.dwell_time)
-        last = Signal()
-        m.d.comb += [
-            self.super_dac_stream.payload.dac_x_code.eq(self.dac_stream_data.dac_x_code),
-            self.super_dac_stream.payload.dac_y_code.eq(self.dac_stream_data.dac_y_code),
-            self.super_dac_stream.payload.blank.eq(self.dac_stream_data.blank),
-            self.super_dac_stream.payload.delay.eq(self.dac_stream_data.delay),
-            last.eq(dwell_counter == self.dac_stream_data.dwell_time)
-            #self.super_dac_stream.payload.last.eq(dwell_counter == self.dac_stream_data.dwell_time),
-        ]
-        with m.If(self.stall_count_reset):
-            m.d.sync += self.stall_cycles.eq(0)
-
-        stalled = Signal()
-        with m.FSM():
-            with m.State("Wait"):
-                m.d.comb += self.dac_stream.ready.eq(1)
-                with m.If(self.dac_stream.valid):
-                    m.d.sync += self.dac_stream_data.eq(self.dac_stream.payload)
-                    m.d.sync += dwell_counter.eq(0)
-                    # m.d.sync += delay_counter.eq(0)
-                    m.next = "Generate"
-                
-
-            with m.State("Generate"):
-                m.d.comb += self.super_dac_stream.valid.eq(1)
-                with m.If(self.super_dac_stream.ready):
-                    with m.If(last):
-                        m.d.comb += self.super_dac_stream.payload.last.eq(last)
-                        m.next = "Wait"
-                    with m.Else():
-                        m.d.sync += dwell_counter.eq(dwell_counter + 1)
-
-        m.d.comb += self.encoder.i.eq(sample_counter)
-        running_sum = Signal(30)
-        last_p2_sum = Signal(30)
-        selected_sum = Signal(30)
-        shifted_sum = Signal(30)
-        with m.If(self.encoder.p): #if the current sample counter is a power of 2, use all samples
-            m.d.comb += selected_sum.eq(running_sum)
-            m.d.sync += last_p2_sum.eq(running_sum)
-        with m.Else(): # else, only average up to the last power of 2 samples
-            m.d.comb += selected_sum.eq(last_p2_sum)
-        m.d.comb += self.adc_stream.payload.adc_code.eq(selected_sum >> self.encoder.o)
-
-        with m.FSM():
-            with m.State("Start"):
-                m.d.comb += self.super_adc_stream.ready.eq(1)
-                with m.If(self.super_adc_stream.valid):
-                    m.d.sync += sample_counter.eq(1)
-                    m.d.sync += running_sum.eq(self.super_adc_stream.payload.adc_code)
-                    with m.If(self.super_adc_stream.payload.last):
-                        m.next = "Wait"
-                    with m.Else():
-                        m.next = "Average"
-
-            with m.State("Average"):
-                m.d.comb += self.super_adc_stream.ready.eq(1)
-                with m.If(self.super_adc_stream.valid):
-                    m.d.sync += sample_counter.eq(sample_counter + 1)
-                    m.d.sync += running_sum.eq(running_sum + self.super_adc_stream.payload.adc_code)
-                    with m.If(self.super_adc_stream.payload.last):
-                        m.next = "Wait"
-                    with m.Else():
-                        m.next = "Average"
-
-
-            with m.State("Wait"):
-                m.d.comb += self.adc_stream.valid.eq(1)
-                with m.If(self.adc_stream.ready):
-                    m.next = "Start"
-
-        return m
-
-#=========================================================================
-
-class RasterRegion(data.Struct):
-    x_start: 14 # UQ(14,0)
-    padding_x_start: 2
-    x_count: 14 # UQ(14,0)
-    padding_x_count: 2
-    x_step:  16 # UQ(8,8)
-    y_start: 14 # UQ(14,0)
-    padding_y_start: 2
-    y_count: 14 # UQ(14,0)
-    padding_y_count: 2
-    y_step:  16 # UQ(8,8)
-
-
-class RasterScanner(wiring.Component):
-    """
-    Properties:
-        FRAC_BITS: number of fixed fractional bits in accumulators
-
-    In:
-        roi_stream: A RasterRegion provided by a RasterScanCommand
-        dwell_stream: A dwell time value provided by one of the RasterPixel commands
-        abort: Interrupt the scan in progress and fetch the next ROI from `roi_stream`
-    Out:
-        dac_stream: X and Y DAC codes and a dwell time
-    """
-    FRAC_BITS = 8
-
-    roi_stream: In(StreamSignature(RasterRegion))
-
-    dwell_stream: In(StreamSignature(data.StructLayout({
-        "dwell_time": DwellTime,
-        "blank": BlankRequest,
-    })))
-
-    abort: In(1)
-    #: Interrupt the scan in progress and fetch the next ROI from `roi_stream`.
-
-    dac_stream: Out(StreamSignature(DACStream))
-
-    def elaborate(self, platform):
-        m = Module()
-
-        region  = Signal.like(self.roi_stream.payload)
-
-        x_accum = Signal(14 + self.FRAC_BITS)
-        x_count = Signal.like(region.x_count)
-        y_accum = Signal(14 + self.FRAC_BITS)
-        y_count = Signal.like(region.y_count)
-        m.d.comb += [
-            self.dac_stream.payload.dac_x_code.eq(x_accum >> self.FRAC_BITS),
-            self.dac_stream.payload.dac_y_code.eq(y_accum >> self.FRAC_BITS),
-            self.dac_stream.payload.dwell_time.eq(self.dwell_stream.payload.dwell_time),
-            self.dac_stream.payload.blank.eq(self.dwell_stream.payload.blank)
-        ]
-
-        with m.FSM():
-            with m.State("Get-ROI"):
-                m.d.comb += self.roi_stream.ready.eq(1)
-                with m.If(self.roi_stream.valid):
-                    m.d.sync += [
-                        region.eq(self.roi_stream.payload),
-                        x_accum.eq(self.roi_stream.payload.x_start << self.FRAC_BITS),
-                        x_count.eq(self.roi_stream.payload.x_count - 1),
-                        y_accum.eq(self.roi_stream.payload.y_start << self.FRAC_BITS),
-                        y_count.eq(self.roi_stream.payload.y_count - 1),
-                    ]
-                    m.next = "Scan"
-
-            with m.State("Scan"):
-                m.d.comb += self.dwell_stream.ready.eq(self.dac_stream.ready)
-                m.d.comb += self.dac_stream.valid.eq(self.dwell_stream.valid)
-                with m.If(self.dac_stream.ready):
-                    with m.If(self.abort):
-                        m.next = "Get-ROI"
-                with m.If(self.dwell_stream.valid & self.dac_stream.ready):
-                    # AXI4-Stream §2.2.1
-                    # > Once TVALID is asserted it must remain asserted until the handshake occurs.
-
-                    ## TODO: AC line sync 
-                    ## TODO: external trigger
-                    ## TODO: be flyback aware, line and frame
-
-                    with m.If(x_count == 0):
-                        with m.If(y_count == 0):
-                            m.next = "Get-ROI"
-                        with m.Else():
-                            m.d.sync += y_accum.eq(y_accum + region.y_step)
-                            m.d.sync += y_count.eq(y_count - 1)
-
-                        m.d.sync += x_accum.eq(region.x_start << self.FRAC_BITS)
-                        m.d.sync += x_count.eq(region.x_count - 1)
-                    with m.Else():
-                        m.d.sync += x_accum.eq(x_accum + region.x_step)
-                        m.d.sync += x_count.eq(x_count - 1)
-            
-        return m
-
-#=========================================================================
-
-class CommandParser(wiring.Component):
-    usb_stream: In(StreamSignature(8))
-    cmd_stream: Out(StreamSignature(Command))
-
-    def elaborate(self, platform):
-        m = Module()
-        self.command = Signal(Command)
-        m.d.comb += self.cmd_stream.payload.eq(self.command)
-        self.command_reg = Signal(Command)
-        array_length = Signal(16)
-
-        self.is_started = Signal()
-        with m.FSM() as fsm:
-            m.d.comb += self.is_started.eq(fsm.ongoing("Type"))
-            def goto_first_deserialized_state(from_type=self.command.type):
-                with m.Switch(from_type):
-                    for cmdtype, state_sequence in Command.deserialized_states.items():
-                        with m.Case(cmdtype):
-                            if len(state_sequence.keys()) > 0:
-                                m.next = list(state_sequence.keys())[0]
-                            else:
-                                m.next = "Submit"
-
-            with m.State("Type"):
-                m.d.comb += self.usb_stream.ready.eq(1)
-                with m.If(self.usb_stream.valid):
-                    m.d.comb += self.command.type.eq(self.usb_stream.payload[4:8])
-                    m.d.comb += self.command.payload.as_value()[0:4].eq(self.usb_stream.payload[0:4])
-                    m.d.sync += self.command_reg.eq(self.command)
-                    goto_first_deserialized_state()
-                    
-
-            def Deserialize(target, state, next_state):
-                m.d.comb += self.command.eq(self.command_reg)
-                #print(f'state: {state} -> next state: {next_state}')
-                with m.State(state):
-                    m.d.comb += self.usb_stream.ready.eq(1)
-                    with m.If(self.usb_stream.valid):
-                        m.d.sync += target.eq(self.usb_stream.payload)
-                        m.next = next_state
-            
-            for state_sequence in Command.deserialized_states.values():
-                for n, (state, offset) in enumerate(state_sequence.items()):
-                    if n < len(state_sequence) - 1:
-                        next_state = list(state_sequence.keys())[n+1]
-                    elif n == len(state_sequence) - 1:
-                        next_state = "Submit"
-                    Deserialize(self.command_reg.as_value()[offset:offset+8], state, next_state)
-
-
-            with m.State("Submit"):
-                m.d.comb += self.command.eq(self.command_reg)
-                with m.If(self.command.type == CmdType.Array):
-                        m.d.sync += self.command_reg.type.eq(self.command.payload.array.cmdtype)
-                        m.d.sync += self.command_reg.as_value()[4:].eq(0)
-                        m.d.sync += array_length.eq(self.command.payload.array.array_length)
-                        goto_first_deserialized_state(from_type=self.command.payload.array.cmdtype)
-                with m.Else():
-                    with m.If(self.cmd_stream.ready):
-                        m.d.comb += self.cmd_stream.valid.eq(1)
-                        with m.If(array_length != 0):
-                            m.d.sync += array_length.eq(array_length - 1)
-                            goto_first_deserialized_state()
-                        with m.Else():
-                            m.next = "Type"
-        return m
-
-#=========================================================================
 
 class CommandExecutor(wiring.Component):
     cmd_stream: In(StreamSignature(Command))
