+++ software/obi/applet/open_beam_interface/modules/debug.py
@@ -1,0 +1,31 @@
+from amaranth import *
+
+from amaranth.lib import data, wiring
+from amaranth.lib.wiring import In, Out, flipped
+
+from obi.applet.open_beam_interface.modules.structs import BusSignature
+
+class PipelinedLoopbackAdapter(wiring.Component):
+    loopback_stream: In(unsigned(14))
+    bus: Out(BusSignature)
+
+    def __init__(self, adc_latency: int):
+        self.adc_latency = adc_latency
+        super().__init__()
+
+    def elaborate(self, platform):
+        m = Module()
+
+        prev_bus_adc_oe = Signal()
+        adc_oe_falling = Signal()
+        m.d.sync += prev_bus_adc_oe.eq(self.bus.adc_oe)
+        m.d.comb += adc_oe_falling.eq(prev_bus_adc_oe & ~self.bus.adc_oe)
+
+        shift_register = Signal(14*self.adc_latency)
+
+        with m.If(adc_oe_falling):
+            m.d.sync += shift_register.eq((shift_register << 14) | self.loopback_stream)
+
+        m.d.comb += self.bus.data_i.eq(shift_register.word_select(self.adc_latency-1, 14))
+
+        return m
